+++ docs/_build/html/configuration.php
@@ -5,7 +5,7 @@
   <head>
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
     
-    <title>Configuration &mdash; PyiUpdater 0.13.1-dev111516 documentation</title>
+    <title>Configuration &mdash; PyiUpdater 0.14-dev141210 documentation</title>
     
     <link rel="stylesheet" href="_static/basic.css" type="text/css" />
     <link rel="stylesheet" href="_static/pygments.css" type="text/css" />
@@ -15,9 +15,9 @@
     <script type="text/javascript">
       var DOCUMENTATION_OPTIONS = {
         URL_ROOT:    './',
-        VERSION:     '0.13.1-dev111516',
+        VERSION:     '0.14-dev141210',
         COLLAPSE_INDEX: false,
-        FILE_SUFFIX: '.html',
+        FILE_SUFFIX: '.php',
         HAS_SOURCE:  true
       };
     </script>
@@ -28,9 +28,9 @@
     <script type="text/javascript" src="_static/js/jquery-fix.js"></script>
     <script type="text/javascript" src="_static/bootstrap-3.2.0/js/bootstrap.min.js"></script>
     <script type="text/javascript" src="_static/bootstrap-sphinx.js"></script>
-    <link rel="top" title="PyiUpdater 0.13.1-dev111516 documentation" href="index.html" />
-    <link rel="next" title="Usage" href="usage.html" />
-    <link rel="prev" title="Installation" href="installation.html" />
+    <link rel="top" title="PyiUpdater 0.14-dev141210 documentation" href="index.php" />
+    <link rel="next" title="Usage" href="usage.php" />
+    <link rel="prev" title="Installation" href="installation.php" />
 <meta charset='utf-8'>
 <meta http-equiv='X-UA-Compatible' content='IE=edge,chrome=1'>
 <meta name='viewport' content='width=device-width, initial-scale=1.0, maximum-scale=1'>
@@ -48,9 +48,9 @@
           <span class="icon-bar"></span>
           <span class="icon-bar"></span>
         </button>
-        <a class="navbar-brand" href="index.html">
+        <a class="navbar-brand" href="index.php">
           PyiUpdater</a>
-        <span class="navbar-text navbar-version pull-left"><b>0.13.1-dev111516</b></span>
+        <span class="navbar-text navbar-version pull-left"><b>0.14-dev141210</b></span>
       </div>
 
         <div class="collapse navbar-collapse nav-collapse">
@@ -62,20 +62,22 @@
      id="dLabelGlobalToc"
      data-toggle="dropdown"
      data-target="#"
-     href="index.html">Site <b class="caret"></b></a>
+     href="index.php">Site <b class="caret"></b></a>
   <ul class="dropdown-menu globaltoc"
       role="menu"
       aria-labelledby="dLabelGlobalToc"><ul class="current">
-<li class="toctree-l1"><a class="reference internal" href="downloads.html">Downloads</a></li>
-<li class="toctree-l1"><a class="reference internal" href="installation.html">Installation</a></li>
+<li class="toctree-l1"><a class="reference internal" href="downloads.php">Downloads</a></li>
+<li class="toctree-l1"><a class="reference internal" href="installation.php">Installation</a></li>
 <li class="toctree-l1 current"><a class="current reference internal" href="">Configuration</a></li>
-<li class="toctree-l1"><a class="reference internal" href="usage.html">Usage</a></li>
-<li class="toctree-l1"><a class="reference internal" href="architecture.html">Architecture</a></li>
-<li class="toctree-l1"><a class="reference internal" href="folder_structure.html">File &amp; Folder Structure</a></li>
-<li class="toctree-l1"><a class="reference internal" href="contributing.html">Contributing</a></li>
-<li class="toctree-l1"><a class="reference internal" href="api.html">API</a></li>
-<li class="toctree-l1"><a class="reference internal" href="license.html">License</a></li>
-<li class="toctree-l1"><a class="reference internal" href="release_history.html">Release History</a></li>
+<li class="toctree-l1"><a class="reference internal" href="usage.php">Usage</a></li>
+<li class="toctree-l1"><a class="reference internal" href="architecture.php">Architecture</a></li>
+<li class="toctree-l1"><a class="reference internal" href="folder_structure.php">File &amp; Folder Structure</a></li>
+<li class="toctree-l1"><a class="reference internal" href="contributing.php">Contributing</a></li>
+<li class="toctree-l1"><a class="reference internal" href="api.php">API</a></li>
+<li class="toctree-l1"><a class="reference internal" href="license.php">License</a></li>
+<li class="toctree-l1"><a class="reference internal" href="release_history.php">Release History</a></li>
+<li class="toctree-l1"><a class="reference internal" href="release_history.php#if-you-update-to-this-release-do-not-revoke-any-keys-until-you-are-sure-all-clients-are-updated-to-this-version-of-the-framework-if-you-revoke-a-key-it-will-break-the-built-in-migration">If you update to this release, do not revoke any keys until you are sure all clients are updated to this version of the framework. If you revoke a key it will break the built in migration.</a></li>
+<li class="toctree-l1"><a class="reference internal" href="release_history.php#demos-have-been-update-with-the-changes-also-its-very-important-to-make-a-decrypted-copy-of-your-config-file-before-updating">Demos have been update with the changes. Also its very important to make a decrypted copy of your config file before updating.</a></li>
 </ul>
 </ul>
 </li>
@@ -99,11 +101,11 @@
               
                 
   <li>
-    <a href="installation.html" title="Previous Chapter: Installation"><span class="glyphicon glyphicon-chevron-left visible-sm"></span><span class="hidden-sm hidden-tablet">&laquo; Installation</span>
+    <a href="installation.php" title="Previous Chapter: Installation"><span class="glyphicon glyphicon-chevron-left visible-sm"></span><span class="hidden-sm hidden-tablet">&laquo; Installation</span>
     </a>
   </li>
   <li>
-    <a href="usage.html" title="Next Chapter: Usage"><span class="glyphicon glyphicon-chevron-right visible-sm"></span><span class="hidden-sm hidden-tablet">Usage &raquo;</span>
+    <a href="usage.php" title="Next Chapter: Usage"><span class="glyphicon glyphicon-chevron-right visible-sm"></span><span class="hidden-sm hidden-tablet">Usage &raquo;</span>
     </a>
   </li>
               
@@ -121,7 +123,7 @@
 
           
             
-<form class="navbar-form navbar-right" action="search.html" method="get">
+<form class="navbar-form navbar-right" action="search.php" method="get">
  <div class="form-group">
   <input type="text" name="q" class="form-control" placeholder="Search" />
  </div>
@@ -203,7 +205,7 @@ uploads</td>
     </p>
     <p>
         &copy; Copyright 2014, Digital Sapphire.<br/>
-      Last updated on Jan 01, 2015.<br/>
+      Last updated on Jan 04, 2015.<br/>
     </p>
   </div>
 </footer>