+++ docs/_build/html/py-modindex.php
@@ -5,7 +5,7 @@
   <head>
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
     
-    <title>Python Module Index &mdash; PyiUpdater 0.13.1-dev111516 documentation</title>
+    <title>Python Module Index &mdash; PyiUpdater 0.14-dev141210 documentation</title>
     
     <link rel="stylesheet" href="_static/basic.css" type="text/css" />
     <link rel="stylesheet" href="_static/pygments.css" type="text/css" />
@@ -15,9 +15,9 @@
     <script type="text/javascript">
       var DOCUMENTATION_OPTIONS = {
         URL_ROOT:    './',
-        VERSION:     '0.13.1-dev111516',
+        VERSION:     '0.14-dev141210',
         COLLAPSE_INDEX: false,
-        FILE_SUFFIX: '.html',
+        FILE_SUFFIX: '.php',
         HAS_SOURCE:  true
       };
     </script>
@@ -28,7 +28,7 @@
     <script type="text/javascript" src="_static/js/jquery-fix.js"></script>
     <script type="text/javascript" src="_static/bootstrap-3.2.0/js/bootstrap.min.js"></script>
     <script type="text/javascript" src="_static/bootstrap-sphinx.js"></script>
-    <link rel="top" title="PyiUpdater 0.13.1-dev111516 documentation" href="index.html" />
+    <link rel="top" title="PyiUpdater 0.14-dev141210 documentation" href="index.php" />
 
 <meta charset='utf-8'>
 <meta http-equiv='X-UA-Compatible' content='IE=edge,chrome=1'>
@@ -49,9 +49,9 @@
           <span class="icon-bar"></span>
           <span class="icon-bar"></span>
         </button>
-        <a class="navbar-brand" href="index.html">
+        <a class="navbar-brand" href="index.php">
           PyiUpdater</a>
-        <span class="navbar-text navbar-version pull-left"><b>0.13.1-dev111516</b></span>
+        <span class="navbar-text navbar-version pull-left"><b>0.14-dev141210</b></span>
       </div>
 
         <div class="collapse navbar-collapse nav-collapse">
@@ -63,20 +63,22 @@
      id="dLabelGlobalToc"
      data-toggle="dropdown"
      data-target="#"
-     href="index.html">Site <b class="caret"></b></a>
+     href="index.php">Site <b class="caret"></b></a>
   <ul class="dropdown-menu globaltoc"
       role="menu"
       aria-labelledby="dLabelGlobalToc"><ul>
-<li class="toctree-l1"><a class="reference internal" href="downloads.html">Downloads</a></li>
-<li class="toctree-l1"><a class="reference internal" href="installation.html">Installation</a></li>
-<li class="toctree-l1"><a class="reference internal" href="configuration.html">Configuration</a></li>
-<li class="toctree-l1"><a class="reference internal" href="usage.html">Usage</a></li>
-<li class="toctree-l1"><a class="reference internal" href="architecture.html">Architecture</a></li>
-<li class="toctree-l1"><a class="reference internal" href="folder_structure.html">File &amp; Folder Structure</a></li>
-<li class="toctree-l1"><a class="reference internal" href="contributing.html">Contributing</a></li>
-<li class="toctree-l1"><a class="reference internal" href="api.html">API</a></li>
-<li class="toctree-l1"><a class="reference internal" href="license.html">License</a></li>
-<li class="toctree-l1"><a class="reference internal" href="release_history.html">Release History</a></li>
+<li class="toctree-l1"><a class="reference internal" href="downloads.php">Downloads</a></li>
+<li class="toctree-l1"><a class="reference internal" href="installation.php">Installation</a></li>
+<li class="toctree-l1"><a class="reference internal" href="configuration.php">Configuration</a></li>
+<li class="toctree-l1"><a class="reference internal" href="usage.php">Usage</a></li>
+<li class="toctree-l1"><a class="reference internal" href="architecture.php">Architecture</a></li>
+<li class="toctree-l1"><a class="reference internal" href="folder_structure.php">File &amp; Folder Structure</a></li>
+<li class="toctree-l1"><a class="reference internal" href="contributing.php">Contributing</a></li>
+<li class="toctree-l1"><a class="reference internal" href="api.php">API</a></li>
+<li class="toctree-l1"><a class="reference internal" href="license.php">License</a></li>
+<li class="toctree-l1"><a class="reference internal" href="release_history.php">Release History</a></li>
+<li class="toctree-l1"><a class="reference internal" href="release_history.php#if-you-update-to-this-release-do-not-revoke-any-keys-until-you-are-sure-all-clients-are-updated-to-this-version-of-the-framework-if-you-revoke-a-key-it-will-break-the-built-in-migration">If you update to this release, do not revoke any keys until you are sure all clients are updated to this version of the framework. If you revoke a key it will break the built in migration.</a></li>
+<li class="toctree-l1"><a class="reference internal" href="release_history.php#demos-have-been-update-with-the-changes-also-its-very-important-to-make-a-decrypted-copy-of-your-config-file-before-updating">Demos have been update with the changes. Also its very important to make a decrypted copy of your config file before updating.</a></li>
 </ul>
 </ul>
 </li>
@@ -107,7 +109,7 @@
 
           
             
-<form class="navbar-form navbar-right" action="search.html" method="get">
+<form class="navbar-form navbar-right" action="search.php" method="get">
  <div class="form-group">
   <input type="text" name="q" class="form-control" placeholder="Search" />
  </div>
@@ -138,72 +140,72 @@
        <td><img src="_static/minus.png" class="toggler"
               id="toggle-1" style="display: none" alt="-" /></td>
        <td>
-       <a href="api.html#module-pyi_updater"><tt class="xref">pyi_updater</tt></a></td><td>
+       <a href="api.php#module-pyi_updater"><tt class="xref">pyi_updater</tt></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
        <td>&nbsp;&nbsp;&nbsp;
-       <a href="api.html#module-pyi_updater.client"><tt class="xref">pyi_updater.client</tt></a></td><td>
+       <a href="api.php#module-pyi_updater.client"><tt class="xref">pyi_updater.client</tt></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
        <td>&nbsp;&nbsp;&nbsp;
-       <a href="api.html#module-pyi_updater.client.updates"><tt class="xref">pyi_updater.client.updates</tt></a></td><td>
+       <a href="api.php#module-pyi_updater.client.updates"><tt class="xref">pyi_updater.client.updates</tt></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
        <td>&nbsp;&nbsp;&nbsp;
-       <a href="api.html#module-pyi_updater.client.utils"><tt class="xref">pyi_updater.client.utils</tt></a></td><td>
+       <a href="api.php#module-pyi_updater.client.utils"><tt class="xref">pyi_updater.client.utils</tt></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
        <td>&nbsp;&nbsp;&nbsp;
-       <a href="api.html#module-pyi_updater.config"><tt class="xref">pyi_updater.config</tt></a></td><td>
+       <a href="api.php#module-pyi_updater.config"><tt class="xref">pyi_updater.config</tt></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
        <td>&nbsp;&nbsp;&nbsp;
-       <a href="api.html#module-pyi_updater.downloader"><tt class="xref">pyi_updater.downloader</tt></a></td><td>
+       <a href="api.php#module-pyi_updater.downloader"><tt class="xref">pyi_updater.downloader</tt></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
        <td>&nbsp;&nbsp;&nbsp;
-       <a href="api.html#module-pyi_updater.exceptions"><tt class="xref">pyi_updater.exceptions</tt></a></td><td>
+       <a href="api.php#module-pyi_updater.exceptions"><tt class="xref">pyi_updater.exceptions</tt></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
        <td>&nbsp;&nbsp;&nbsp;
-       <a href="api.html#module-pyi_updater.key_handler"><tt class="xref">pyi_updater.key_handler</tt></a></td><td>
+       <a href="api.php#module-pyi_updater.key_handler"><tt class="xref">pyi_updater.key_handler</tt></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
        <td>&nbsp;&nbsp;&nbsp;
-       <a href="api.html#module-pyi_updater.package_handler"><tt class="xref">pyi_updater.package_handler</tt></a></td><td>
+       <a href="api.php#module-pyi_updater.package_handler"><tt class="xref">pyi_updater.package_handler</tt></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
        <td>&nbsp;&nbsp;&nbsp;
-       <a href="api.html#module-pyi_updater.package_handler.package"><tt class="xref">pyi_updater.package_handler.package</tt></a></td><td>
+       <a href="api.php#module-pyi_updater.package_handler.package"><tt class="xref">pyi_updater.package_handler.package</tt></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
        <td>&nbsp;&nbsp;&nbsp;
-       <a href="api.html#module-pyi_updater.patcher"><tt class="xref">pyi_updater.patcher</tt></a></td><td>
+       <a href="api.php#module-pyi_updater.patcher"><tt class="xref">pyi_updater.patcher</tt></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
        <td>&nbsp;&nbsp;&nbsp;
-       <a href="api.html#module-pyi_updater.uploader"><tt class="xref">pyi_updater.uploader</tt></a></td><td>
+       <a href="api.php#module-pyi_updater.uploader"><tt class="xref">pyi_updater.uploader</tt></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
        <td>&nbsp;&nbsp;&nbsp;
-       <a href="api.html#module-pyi_updater.uploader.common"><tt class="xref">pyi_updater.uploader.common</tt></a></td><td>
+       <a href="api.php#module-pyi_updater.uploader.common"><tt class="xref">pyi_updater.uploader.common</tt></a></td><td>
        <em></em></td></tr>
      <tr class="cg-1">
        <td></td>
        <td>&nbsp;&nbsp;&nbsp;
-       <a href="api.html#module-pyi_updater.utils"><tt class="xref">pyi_updater.utils</tt></a></td><td>
+       <a href="api.php#module-pyi_updater.utils"><tt class="xref">pyi_updater.utils</tt></a></td><td>
        <em></em></td></tr>
    </table>
 
@@ -220,7 +222,7 @@
     </p>
     <p>
         &copy; Copyright 2014, Digital Sapphire.<br/>
-      Last updated on Jan 01, 2015.<br/>
+      Last updated on Jan 04, 2015.<br/>
     </p>
   </div>
 </footer>